"eval" results can be browsed and modified (if they are complex types)
Due to a bug on Eclipse side, the Expressions view is not refreshed
immediately, it requires to switch to another tab and get back to force
expressions to re-evaluate.

This is implemented as a cache in a new context (so cached results are
flushed when program execution restart). This context (-1) is invisible
to context_names command and should not be queryed with context_get.